chore: rename demos to gut_study and landmarks
Promote the neighbors notebook to demos/landmarks.py and move the gut playground to demos/gut_study.py.
